# spec/swagger.yaml
components:
  schemas:
    APIKey:
      properties:
        created_at:
          type: string
        id:
          example: key_4n8q2w6e1r5t9y3u
          type: string
        last_used_at:
          type: string
        name:
          type: string
        prefix:
          example: eyf_a1b2c3
          type: string
      required:
      - created_at
      - id
      - name
      - prefix
      type: object
    APIKeyList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/APIKey'
          type: array
          uniqueItems: false
      required:
      - items
      type: object
...
    CreateAPIKeyRequest:
      properties:
        name:
          example: ci
          type: string
      required:
      - name
      type: object
...
    CreatedAPIKey:
      properties:
        key:
          $ref: '#/components/schemas/APIKey'
        secret:
          example: eyf_…
          type: string
      required:
      - key
      - secret
      type: object
...
    ErrorCode:
      enum:
      - bad_request
      - unauthenticated
      - not_found
      - method_not_allowed
      - internal
      - unavailable
      - review_not_found
      - idempotency_conflict
      - rate_limited
      - invalid_credentials
      - email_taken
      - invalid_email
      - weak_password
      - api_key_not_found
      - session_not_found
      - review_running
      - origin_forbidden
      - repository_not_found
      - provider_not_linked
      - provider_expired
      type: string
      x-enum-varnames:
      - CodeBadRequest
      - CodeUnauthenticated
      - CodeNotFound
      - CodeMethodNotAllowed
      - CodeInternal
      - CodeUnavailable
      - CodeReviewNotFound
      - CodeIdempotencyConflict
      - CodeRateLimited
      - CodeInvalidCredentials
      - CodeEmailTaken
      - CodeInvalidEmail
      - CodeWeakPassword
      - CodeAPIKeyNotFound
      - CodeSessionNotFound
      - CodeReviewRunning
      - CodeOriginForbidden
      - CodeRepositoryNotFound
      - CodeProviderNotLinked
      - CodeProviderExpired
...
    OAuthProvider:
      properties:
        id:
          enum:
          - github
          - gitee
          example: github
          type: string
        title:
          example: GitHub
          type: string
      required:
      - id
      - title
      type: object
    OAuthProviderList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/OAuthProvider'
          type: array
          uniqueItems: false
      required:
      - items
      type: object
    Problem:
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          example: review_not_found
        detail:
          example: review r_123 does not exist
          type: string
        request_id:
          type: string
        status:
          example: 404
          type: integer
        title:
          example: Not Found
          type: string
        type:
          example: about:blank
          type: string
      required:
      - code
      - status
      - title
      - type
      type: object
...
    UpdateAPIKeyRequest:
      properties:
        name:
          example: ci-main
          type: string
      required:
      - name
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Browser session set by sign-in
      in: cookie
      name: eyeful_session
      type: apiKey
...
info:
  description: Multi-agent code review with executable evidence.
  title: eyeful API
  version: 0.1.0
openapi: 3.1.0
paths:
  /api-keys:
    get:
      operationId: listAPIKeys
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKeyList'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
      security:
      - SessionCookie: []
      summary: List API keys
      tags:
      - api-keys
    post:
      description: The secret is returned only here; store it now.
      operationId: createAPIKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
              description: Key name
              summary: body
        description: Key name
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedAPIKey'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
      security:
      - SessionCookie: []
      summary: Create an API key
      tags:
      - api-keys
  /api-keys/{id}:
    delete:
      operationId: deleteAPIKey
      parameters:
      - description: Key ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
      security:
      - SessionCookie: []
      summary: Revoke an API key
      tags:
      - api-keys
    get:
      operationId: getAPIKey
      parameters:
      - description: Key ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKey'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
      security:
      - SessionCookie: []
      summary: Get one API key
      tags:
      - api-keys
    patch:
      operationId: updateAPIKey
      parameters:
      - description: Key ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAPIKeyRequest'
              description: New name
              summary: body
        description: New name
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKey'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
      security:
      - SessionCookie: []
      summary: Rename an API key
      tags:
      - api-keys
...
  /oauth/{provider}/authorize:
    get:
      operationId: authorizeOAuth
      parameters:
      - description: Provider
        in: path
        name: provider
        required: true
        schema:
          enum:
          - github
          - gitee
          type: string
      responses:
        "302":
          description: Found
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
      summary: Redirect to the provider to sign in
      tags:
      - oauth
  /oauth/{provider}/callback:
    get:
      operationId: callbackOAuth
      parameters:
      - description: Provider
        in: path
        name: provider
        required: true
        schema:
          enum:
          - github
          - gitee
          type: string
      - description: Authorization code
        in: query
        name: code
        required: true
        schema:
          type: string
      - description: State from the authorize redirect
        in: query
        name: state
        required: true
        schema:
          type: string
      responses:
        "302":
          description: Found
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
      summary: The provider redirects here; sets the session and returns to the console
      tags:
      - oauth
  /oauth/providers:
    get:
      operationId: listOAuthProviders
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthProviderList'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal Server Error
      summary: Sign-in providers this server is configured for
      tags:
      - oauth
...
servers:
- url: /